requirements.txt
pillow
numpy
ffmpeg
pafy
python_vlc
cv2
image_to_Ascii